+++ dblib/PostgreSQL.c
@@ -19,9 +19,9 @@
  */
 
 /*
+*/
 #define	DEBUG
 #define	TRACE
-*/
 
 #ifdef HAVE_CONFIG_H
 #  include <config.h>
@@ -1645,7 +1645,7 @@ ENTER_FUNC;
 	} else {
 		db = rec->opt.db;
 		path = db->path[ctrl->pno];
-		if		(  ( ix = (int)g_hash_table_lookup(path->opHash,name) )  ==  0  ) {
+		if		(  ( ix = (int)(long)g_hash_table_lookup(path->opHash,name) )  ==  0  ) {
 			ctrl->rc = MCP_BAD_FUNC;
 			rc = FALSE;
 		} else {
