+++ libs/front.c
@@ -2,7 +2,7 @@
  * PANDA -- a simple transaction monitor
  * Copyright (C) 1998-1999 Ogochan.
  * Copyright (C) 2000-2003 Ogochan & JMA (Japan Medical Association).
- * Copyright (C) 2004-2006 Ogochan.
+ * Copyright (C) 2004-2007 Ogochan.
  * 
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -74,7 +74,7 @@ ENTER_FUNC;
 	InitPool();
 	scr = NewScreenData();
 	scr->status = APL_SESSION_LINK;
-	rseed = (int)scr;
+	rseed = (int)(long)scr;
 LEAVE_FUNC;
 	return	(scr); 
 }
